In files with `mod x;`, expose uses of `x`.
This is only halfway to working. We aim to make a path from the module's
definition node (in `x.rs`) to its uses in the parent module (the file that
contains `mod x;`), via the root node. This is only the second half of the
path, from the root node to uses in the parent module.